oui bravo

+++ src/ch/epfl/javelo/gui/MapViewParameters.java
@@ -4,17 +4,17 @@ import ch.epfl.javelo.projection.PointWebMercator;
 
 import java.awt.geom.Point2D;
 
-public record MapViewParameters(int zoomLevel, double x, double y) {
+public record MapViewParameters(int zoomLevel, int x, int y) {
 
     public Point2D topLeft(){
         return new Point2D.Double(x,y);
     }
 
-    public MapViewParameters withMinXY(double newX){
+    public MapViewParameters withMinXY(int newX){
         return new MapViewParameters(zoomLevel, newX, y);
     }
 
-    public PointWebMercator pointAt(double x, double y){
+    public PointWebMercator pointAt(int x, int y){
         return new PointWebMercator(x+this.x,y+this.y);
     }
